requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
greenlet==3.1.1
idna==3.10
pyTelegramBotAPI==4.26.0
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.38
typing_extensions==4.12.2
urllib3==2.3.0
